pyproject.toml
[project]
name = "blog-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.6",
    "django-debug-toolbar>=6.0.0",
    "django-filter>=25.1",
    "djangorestframework>=3.16.1",
    "faker>=37.6.0",
    "psycopg2>=2.9.10",
    "requests>=2.32.5",
]
